Consolidate map
